Implemented the Template Method design pattern to create the feed filtering algorithm for the program with the parent class Filter.java and its subclasses FilterByCuisine.java, FilterByFollowing.java, and FilterByLikes.java. The filterFeed() function is overriden from the base parent class based on the type of filter of the subclasses.
The feed filtering process involves filtering the right posts, limiting the number of displayed posts, and sortng the posts from the most recent to the oldest.

Updated Post.java to be able to sort a list of posts based on the number of likes on a post through the use of Comparator.